HIR-003: Complete Node.js extension JSON migration
- Add extension JSON files for: assert, console, crypto, net, perf_hooks, util, zlib
- Each extension defines hirName mappings for camelCase to snake_case function names
- Lowering specifications for proper LLVM argument/return types
- Update test fixtures for HIR pipeline compatibility
- Runtime improvements for BigInt, closures, and regex handling

Node tests: 148/280 (52.9%), 87 compile errors remaining

+++ src/runtime/src/Math.cpp
@@ -5,11 +5,11 @@
 
 extern "C" {
 
-int64_t ts_math_min(int64_t a, int64_t b) {
+double ts_math_min(double a, double b) {
     return std::min(a, b);
 }
 
-int64_t ts_math_max(int64_t a, int64_t b) {
+double ts_math_max(double a, double b) {
     return std::max(a, b);
 }
